--- src/WpStarter/Wordpress/Auth/Access/Authorizable.php
@@ -1,56 +1,0 @@
-<?php
-
-namespace WpStarter\Wordpress\Auth\Access;
-
-use WpStarter\Contracts\Auth\Access\Gate;
-
-trait Authorizable
-{
-    /**
-     * Determine if the entity has the given abilities.
-     *
-     * @param iterable|string $abilities
-     * @param array|mixed $arguments
-     * @return bool
-     */
-    public function can($abilities, $arguments = [])
-    {
-        return ws_app(Gate::class)->forUser($this)->check($abilities, $arguments);
-    }
-
-    /**
-     * Determine if the entity has any of the given abilities.
-     *
-     * @param iterable|string $abilities
-     * @param array|mixed $arguments
-     * @return bool
-     */
-    public function canAny($abilities, $arguments = [])
-    {
-        return ws_app(Gate::class)->forUser($this)->any($abilities, $arguments);
-    }
-
-    /**
-     * Determine if the entity does not have the given abilities.
-     *
-     * @param iterable|string $abilities
-     * @param array|mixed $arguments
-     * @return bool
-     */
-    public function cant($abilities, $arguments = [])
-    {
-        return !$this->can($abilities, $arguments);
-    }
-
-    /**
-     * Determine if the entity does not have the given abilities.
-     *
-     * @param iterable|string $abilities
-     * @param array|mixed $arguments
-     * @return bool
-     */
-    public function cannot($abilities, $arguments = [])
-    {
-        return $this->cant($abilities, $arguments);
-    }
-}
