+++ src/WpStarter/Wordpress/Model/Concerns/HasRelationships.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace WpStarter\Wordpress\Auth\Concerns;
+namespace WpStarter\Wordpress\Model\Concerns;
 
 use Closure;
 use WpStarter\Database\ClassMorphViolationException;
